Move the poll folder
