+++ services/agora/src/components/post/display/poll/PollWrapper.vue
@@ -2,7 +2,7 @@
   <div>
     <div v-if="dataLoaded" class="pollContainer">
       <div class="pollOptionList">
-        <poll-option
+        <PollOption
           v-for="optionItem in localPollOptionList"
           :key="optionItem.index"
           :option="optionItem.option"
@@ -70,8 +70,7 @@
 </template>
 
 <script setup lang="ts">
-import PollOption from "./PollOption.vue";
-import ZKButton from "../../ui-library/ZKButton.vue";
+import ZKButton from "../../../ui-library/ZKButton.vue";
 import {
   useHomeFeedStore,
   type DummyPollOptionFormat,
@@ -80,11 +79,12 @@ import { onBeforeMount, ref, watch } from "vue";
 import { useBackendPollApi } from "src/utils/api/poll";
 import type { UserInteraction, PollList } from "src/shared/types/zod";
 import { storeToRefs } from "pinia";
-import ZKIcon from "../../ui-library/ZKIcon.vue";
+import ZKIcon from "../../../ui-library/ZKIcon.vue";
 import { useLoginIntentionStore } from "src/stores/loginIntention";
-import PreLoginIntentionDialog from "../../authentication/intention/PreLoginIntentionDialog.vue";
+import PreLoginIntentionDialog from "../../../authentication/intention/PreLoginIntentionDialog.vue";
 import { useAuthenticationStore } from "src/stores/authentication";
 import { useBackendAuthApi } from "src/utils/api/auth";
+import PollOption from "./PollOption.vue";
 
 const props = defineProps<{
   userResponse: UserInteraction;
